Restructured DH chain class to remove .hpp
