+++ rct_optimizations/include/rct_optimizations/dh_chain.h
@@ -7,6 +7,9 @@
 
 namespace rct_optimizations
 {
+/**
+ * @brief The joint types for DH representation
+ */
 enum class DHJointType : unsigned
 {
   LINEAR,
@@ -39,19 +42,54 @@ struct DHTransform
   virtual ~DHTransform() = default;
 
   /**
-   *
-   */
-  template<typename T>
-  Isometry3<T> createRelativeTransform(const T joint_value,
-                                       const Eigen::Matrix<T, 1, 4>& offsets) const;
-
-  /**
    * @brief Creates the homogoneous transformation from the previous link to the current link
-   * @param joint_value
+   * @param joint_value - The joint value to apply when caluclating the transform
+   * @param offsets - The DH parameter offsets to apply when calculating the transform
    * @return
    */
   template<typename T>
-  Isometry3<T> createRelativeTransform(const T joint_value) const;
+  Isometry3<T> createRelativeTransform(const T joint_value,
+                                       const Eigen::Matrix<T, 1, 4>& offsets) const
+  {
+    Isometry3<T> transform(Isometry3<T>::Identity());
+
+    // Create an DH parameter vector with offsets ([d, theta, r, alpha]
+    Vector4<T> updated_params = params.cast<T>() + offsets.transpose();
+
+    switch (type)
+    {
+    case DHJointType::LINEAR:
+      // Add the joint value to d (index 0) if the joint is linear
+      updated_params(0) += joint_value;
+      break;
+    case DHJointType::REVOLUTE:
+      // Add the joint value to theta (index 1) if the joint is revolute
+      updated_params(1) += joint_value;
+      break;
+    default:
+      break;
+    }
+
+    // Perform the DH transformations
+    transform *= Eigen::Translation<T, 3>(T(0.0), T(0.0), updated_params(0));
+    transform.rotate(Eigen::AngleAxis<T>(updated_params(1), Vector3<T>::UnitZ()));
+    transform *= Eigen::Translation<T, 3>(updated_params(2), T(0.0), T(0.0));
+    transform.rotate(Eigen::AngleAxis<T>(updated_params(3), Vector3<T>::UnitX()));
+
+    return transform;
+  }
+
+  /**
+   * @brief Creates the homogoneous transformation from the previous link to the current link without applying DH parameter offsets
+   * @param joint_value - The joint value to apply when calculating the transform
+   * @return
+   */
+  template<typename T>
+  Isometry3<T> createRelativeTransform(const T joint_value) const
+  {
+    Eigen::Matrix<T, 1, 4> offsets = Eigen::Matrix<T, 1, 4>::Zero();
+    return createRelativeTransform(joint_value, offsets);
+  }
 
   double createRandomJointValue() const;
 
@@ -83,7 +121,15 @@ public:
    * @throws Exception if the size of joint values is larger than the number of DH transforms in the robot
    */
   template<typename T>
-  Isometry3<T> getFK(const Eigen::Matrix<T, Eigen::Dynamic, 1> &joint_values) const;
+  Isometry3<T> getFK(const Eigen::Matrix<T, Eigen::Dynamic, 1> &joint_values) const
+  {
+    Isometry3<T> transform(Isometry3<T>::Identity());
+    for (Eigen::Index i = 0; i < joint_values.size(); ++i)
+    {
+      transform = transform * transforms_.at(i)->createRelativeTransform(joint_values[i]);
+    }
+    return transform;
+  }
 
   /**
    * @brief Override function of @ref getFK but using a data pointer for easier integration with Ceres
@@ -93,7 +139,16 @@ public:
    */
   template<typename T>
   Isometry3<T> getFK(const Eigen::Matrix<T, Eigen::Dynamic, 1>& joint_values,
-                     const Eigen::Matrix<T, Eigen::Dynamic, 4>& offsets) const;
+                     const Eigen::Matrix<T, Eigen::Dynamic, 4>& offsets) const
+  {
+    Isometry3<T> transform(Isometry3<T>::Identity());
+    for (Eigen::Index i = 0; i < joint_values.size(); ++i)
+    {
+      const Eigen::Matrix<T, 1, 4> &offset = offsets.row(i);
+      transform = transform * transforms_.at(i)->createRelativeTransform(joint_values[i], offset);
+    }
+    return transform;
+  }
 
   /**
    * @brief Creates a random joint pose by choosing a random uniformly distributed joint value for each joint in the chain
@@ -105,33 +160,22 @@ public:
    * @brief Returns the number of degrees of freedom (i.e. DH transforms) of the chain
    * @return
    */
-  inline std::size_t dof() const
-  {
-    return transforms_.size();
-  }
+  std::size_t dof() const;
 
-  inline Eigen::MatrixX4d getDHTable() const
-  {
-    Eigen::MatrixX4d out(dof(), 4);
-    for (std::size_t i = 0; i < transforms_.size(); ++i)
-    {
-      out.row(i) = transforms_[i]->params.transpose();
-    }
-    return out;
-  }
+  /**
+   * @brief Gets a dof x 4 matrix of the DH parameters
+   * @return
+   */
+  Eigen::MatrixX4d getDHTable() const;
 
-  inline std::vector<DHJointType> getJointTypes() const
-  {
-    std::vector<DHJointType> out;
-    out.reserve(transforms_.size());
-    for (const auto &t : transforms_)
-    {
-      out.push_back(t->type);
-    }
-    return out;
-  }
+  /**
+   * @brief Returns a the joint types of the kinematic chain in order
+   * @return
+   */
+  std::vector<DHJointType> getJointTypes() const;
 
 protected:
+  /** @brief The DH transforms that make up the chain */
   std::vector<DHTransform::Ptr> transforms_;
 };
