+++ rct_optimizations/test/dh_chain_kinematic_calibration_utest.cpp
@@ -2,8 +2,8 @@
 #include <rct_optimizations/dh_chain_kinematic_calibration.h>
 #include <rct_optimizations_tests/dh_chain_observation_creator.h>
 #include <rct_optimizations_tests/utilities.h>
-#include <rct_optimizations/impl/dh_chain.hpp>
 #include <rct_optimizations_tests/observation_creator.h>
+
 using namespace rct_optimizations;
 
 class DHChainKinematicCalibration : public ::testing::Test
